+++ test/unit-tests/components/views/rooms/RoomListPanel/RoomListPrimaryFilters-test.tsx
@@ -12,7 +12,6 @@ import userEvent from "@testing-library/user-event";
 import { type RoomListViewState } from "../../../../../../src/components/viewmodels/roomlist/RoomListViewModel";
 import { SecondaryFilters } from "../../../../../../src/components/viewmodels/roomlist/useFilteredRooms";
 import { RoomListPrimaryFilters } from "../../../../../../src/components/views/rooms/RoomListPanel/RoomListPrimaryFilters";
-import { SortOption } from "../../../../../../src/components/viewmodels/roomlist/useSorter";
 import { FilterKey } from "../../../../../../src/stores/room-list-v3/skip-list/filters";
 
 describe("<RoomListPrimaryFilters />", () => {
@@ -31,10 +30,6 @@ describe("<RoomListPrimaryFilters />", () => {
             ],
             activateSecondaryFilter: () => {},
             activeSecondaryFilter: SecondaryFilters.AllActivity,
-            sort: jest.fn(),
-            activeSortOption: SortOption.Activity,
-            shouldShowMessagePreview: false,
-            toggleMessagePreview: jest.fn(),
             activeIndex: undefined,
         };
     });
